type conversion warning

+++ device/selections/lines/muon/src/DiMuonNoIPLine.cu
@@ -93,7 +93,7 @@ __device__ void di_muon_no_ip_line::di_muon_no_ip_line_t::monitor(
     const auto track1 = static_cast<const Allen::Views::Physics::BasicParticle*>(vertex.child(0));
     const auto track2 = static_cast<const Allen::Views::Physics::BasicParticle*>(vertex.child(1));
     if (track1->ip_chi2() < 6 && track2->ip_chi2() < 6) {
-      float q = sqrt(vertex.m() * vertex.m() - 4 * Allen::mMu * Allen::mMu);
+      float q = sqrtf(vertex.m() * vertex.m() - 4 * Allen::mMu * Allen::mMu);
       if (q < parameters.dev_q_bin_boundaries[n_bins]) {
         unsigned bin = binary_search_rightmost(&parameters.dev_q_bin_boundaries[0], n_bins, q);
         parameters.dev_array_prompt_q[bin]++;
